refactor(ESM): move test helper function away from build code
