+++ lib/promisifiedRequests.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var XMLHttpRequest = require('xmlhttprequest').XMLHttpRequest;
-import fs from 'fs';
 import debug from 'debug';
 
 const log = debug('promisifiedRequests');
@@ -40,21 +39,5 @@ export function request(obj) {
     } else {
       request.send();
     }
-  });
-}
-
-export async function readFileAsync(path) {
-  return await readFile(path);
-}
-
-export function readFile(path) {
-  return new Promise(function(resolve, reject) {
-    fs.readFile(path, 'utf8', function(err, data) {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(data);
-      }
-    });
   });
 }
